# Speedup websocket mask using C copied from Tornado codebase