--- a/MANIFEST.in
+++ b/MANIFEST.in
@@ -1,3 +1,4 @@
 include LICENSE
 
 graft websockets/py35
+include websockets/speedups.c