--- a/tox.ini
+++ b/tox.ini
@@ -1,5 +1,11 @@
 [tox]
-envlist = py33,py34,py35,py36,coverage,flake8,isort
+envlist = py33,py34,py35,py36,{py33,py34,py35,py36}-no-ext,coverage,flake8,isort
+
+setenv =
+    {py33,py34,py35,py36}: WEBSOCKETS_EXTENSION=1
+    # tox's parser chokes if all the setenv entries are conditional.
+    DUMMY=dummy
+    {py33,py34,py35,py36}-no-ext: WEBSOCKETS_EXTENSION=0
 
 [testenv]
 deps =