--- a/tox.ini
+++ b/tox.ini
@@ -1,11 +1,11 @@
 [tox]
-envlist = py33,py34,py35,py36,{py33,py34,py35,py36}-no-ext,coverage,flake8,isort
+envlist = py33,py34,py35,py36,{py33,py34,py35,py36,coverage}-no-ext,coverage,flake8,isort
 
 setenv =
     {py33,py34,py35,py36}: WEBSOCKETS_EXTENSION=1
     # tox's parser chokes if all the setenv entries are conditional.
     DUMMY=dummy
-    {py33,py34,py35,py36}-no-ext: WEBSOCKETS_EXTENSION=0
+    {py33,py34,py35,py36,coverage}-no-ext: WEBSOCKETS_EXTENSION=0
 
 [testenv]
 deps =
@@ -13,6 +13,13 @@ deps =
 commands = python -m unittest
 
 [testenv:coverage]
+commands =
+    python -m coverage erase
+    python -m coverage run --branch --source=websockets -m unittest
+    python -m coverage report --fail-under=100
+deps = coverage
+
+[testenv:coverage-no-ext]
 commands =
     python -m coverage erase
     python -m coverage run --branch --source=websockets -m unittest