--- a/websockets/speedups.c
+++ b/websockets/speedups.c
@@ -47,7 +47,6 @@ static PyMethodDef methods[] = {
     {NULL, NULL, 0, NULL}
 };
 
-#if PY_MAJOR_VERSION >= 3
 static struct PyModuleDef speedupsmodule = {
    PyModuleDef_HEAD_INIT,
    "speedups",
@@ -60,9 +59,3 @@ PyMODINIT_FUNC
 PyInit_speedups(void) {
     return PyModule_Create(&speedupsmodule);
 }
-#else  // Python 2.x
-PyMODINIT_FUNC
-initspeedups(void) {
-    Py_InitModule("websockets.speedups", methods);
-}
-#endif