… websockets/speedups.pyi