Is there anyway to do a benchmark?

Is there a license issue with the Tornado codes?