@RemiCardona 2017-04-12
Personally, I'd be _extremely_ weary of adding C extensions to this project without:
* fully measuring the impact using benchmarks
* making sure the extension stays optional (much like asyncio's own C extensions in 3.6) and that tests are adapted to exercise **all** code paths.

That being said, I did do a few micro benchmarks around the masking operation.
```python
from itertools import cycle

data = b'abcdef' * 10000
mask_bits = b'L\xde\xcb:'  # obtained through struct.pack('!I', random.getrandbits(32))

# original version
masked_data = bytes(b ^ mask_bits[i % 4] for i, b in enumerate(data))

# new version
masked_data2 = bytes(b ^ m for b, m in zip(data, cycle(mask_bits)))
```

Using ipython's `%timeit`, here are some results:
* 3.4: 10.7ms to 6.39ms → 40% speedup
* 3.6: 9.65ms to 5.91ms → 38% speedup

However, with `data = b'abcdef'`, timeit reports 

> The slowest run took 4.65 times longer than the fastest. This could mean that an intermediate result is being cached.

whereas the original version looks much more stable.

So this looks promising, but I have no idea what the impact of this is on the overall performance of websockets. @messense the same applies to your PR, we need to be able to gauge whether adding a C extension to a pure-python project is worth the hassle.

Cheers