@RemiCardona 2017-04-12
Addendum: with really short data, both versions are equally unstable on 3.6, so 🤷‍♀️