@messense 2017-04-12
with `data = b'abcdef'`

```python
In [22]: %timeit pure_python_mask1(mask_bits, data)
100000 loops, best of 3: 2.22 µs per loop

In [23]: %timeit pure_python_mask2(mask_bits, data)
100000 loops, best of 3: 2.1 µs per loop

In [24]: %timeit websocket_mask(mask_bits, data)
The slowest run took 8.93 times longer than the fastest. This could mean that an intermediate result is being cached.
1000000 loops, best of 3: 235 ns per loop
```