@aaugustin 2017-04-13
Thanks for the pull request.

I'm in favor of shipping a C extension to accelerate low-level operations.

One way to figure out the effect is to run the Autohahn benchmark. Instructions for running it are included in the `compliance` directory.

I'm uncomfortable all the pieces that appear to be copy-pasted from other projects and that don't necessarily serve a purpose here. `setup.py` must remain trivial.

A prerequisite for merging this PR is to figure out how to measure the coverage of C code, like Python code. That would have caught the useless conditional branch for Python 3 vs Python 2 that I saw when I skimmed the diff.

There must be a way to run tests with the speedups and without the speedups. The combination of the two runs must have 100% branch coverage of the C and the Python code. This [toy project](https://github.com/aaugustin/sudoku) which I wrote to learn about C extension can provide some hints in this area.

The C module itself is fairly short. I'll read it when I have more time. I'll have to check that it doesn't leak memory (in the absence or presence of errors).