@aaugustin 2017-07-22
Thanks messense for getting the ball rolling on this. I'm a bit more optimistic about #211 which doesn't contain code copyrighted by Facebook, under a license that isn't fully compatible with websockets's license.

I'm closing in favor of #211, but comparing the two PRs and suggesting improvements in #211 based on ideas found in this one.