@aaugustin 2017-07-23
@messense - I didn't end up merging your version, but I'm making you a collaborator to this repo to recognize your contribution there!